requirements.txt
numpy>=1.18.1,<1.19.0
attrs==19.3
cattrs==1.0.0rc
jsonpickle==1.2
networkx
tensorflow==2.3.1
h5py==2.10.0
python-rapidjson
opencv-python-headless==4.2.0.34
pandas
psutil
PySide2==5.14.1
pyzmq
pyyaml
imgaug==0.3.0
scipy<=1.4.1
scikit-image
scikit-learn
scikit-video
imgstore==0.2.9
qimage2ndarray==1.8
jsmin
seaborn
pykalman==0.9.5
segmentation-models==1.0.1
rich==9.10.0